RAINCATCH-420 Create a subscriber to session validation topic in the raincatcher-user module to call the service validate session
